timthumb added

+++ common/models/ImageUpload.php
@@ -31,11 +31,11 @@ class ImageUpload extends Model{
 
     public function getFolder($obj_id = null)
     {
-        if($this->scenario == self::GALLERY_UPLOAD_SCENARIO){
+        if($this->scenario == self::GALLERY_UPLOAD_SCENARIO ){
             $gallery = Gallery::findOne($obj_id);
             $path_to_folder = Yii::getAlias('@backend') . '/web/elfinder/global/gallery/'.$gallery->dir_name."/";
         }elseif($this->scenario == self::ARTICLE_UPLOAD_SCENARIO) {
-            $path_to_folder = Yii::getAlias('@backend') . '/web/elfinder/global/article_'.Yii::$app->request->get('id');
+            $path_to_folder = Yii::getAlias('@backend') . '/web/elfinder/global/article_'.$obj_id;
         }
         if(!is_dir($path_to_folder)){
            mkdir($path_to_folder, 0777);
@@ -48,31 +48,39 @@ class ImageUpload extends Model{
         return strtolower(md5(uniqid($this->image->baseName)) . '.' . $this->image->extension);
     }
 
-    public function deleteCurrentImage($currentImage)
+    public function deleteCurrentImage($currentImage,$obj_id = null)
     {
-        if($this->fileExists($currentImage))
+        if($this->fileExists($currentImage,$obj_id))
         {
-            unlink($this->getFolder() . $currentImage);
+            unlink($this->getFolder($obj_id) . $currentImage);
         }
     }
 
-    public function fileExists($currentImage)
+    public function fileExists($currentImage,$obj_id = null)
     {
         if(!empty($currentImage) && $currentImage != null)
         {
-            return file_exists($this->getFolder() . $currentImage);
+            return file_exists($this->getFolder($obj_id) . $currentImage);
         }
     }
 
     public function saveImage($obj_id = null)
     {
         $image = new Image();
-        $image->name = $this->image->name;
-        $image->gallery_id = $obj_id;
+        if($this->scenario == self::GALLERY_UPLOAD_SCENARIO) {
+            $image->gallery_id = $obj_id;
+            $filename = $this->generateFilename();
+            $image->name = $filename;
+
+        }elseif($this->scenario == self::ARTICLE_UPLOAD_SCENARIO) {
+            $image->name = 'main.jpg';
+            $filename = $image->name;
+        }
+
         if(!$image->save()){
             Functions::pretty_var_dump($image->errors);die();
         }
-        $filename = $this->generateFilename();
+
         $this->image->saveAs($this->getFolder($obj_id) . $filename);
         return $filename;
     }
@@ -83,7 +91,9 @@ class ImageUpload extends Model{
                 is_dir($obj) ? $this->removeDirectory($obj) : unlink($obj);
             }
         }
-        rmdir($dir);
+        if(file_exists($dir)) {
+            rmdir($dir);
+        }
     }
   public function createFolder($article_id)
     {
